backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
pillow
transformers
torch
aiofiles
numpy